# Bundle and minify vendor files