--- a/js/bower.json
+++ b/js/bower.json
@@ -6,7 +6,6 @@
     "jquery-timepicker": "883bb2cd94",
     "jstzdetect": "1.0.6",
     "ical.js": "1.2.2",
-    "davclient.js": "evert/davclient.js",
     "fullcalendar": "3.1.0",
     "hsl_rgb_converter": "https://github.com/kayellpeee/hsl_rgb_converter.git",
     "autosize": "^3.0.20"