--- a/js/licenses/hsl_rgb_converter.js
+++ b/js/licenses/hsl_rgb_converter.js
@@ -1,0 +1,5 @@
+/*!
+ HSL RGB Converter
+ author: Keenan Lidral-Porter
+ license: ISC
+ */