--- a/js/licenses/jquery.timepicker.css
+++ b/js/licenses/jquery.timepicker.css
@@ -1,0 +1,9 @@
+/*
+ * jQuery UI Timepicker
+ *
+ * Copyright 2010-2013, Francois Gelinas
+ * Dual licensed under the MIT or GPL Version 2 licenses.
+ * http://jquery.org/license
+ *
+ * http://fgelinas.com/code/timepicker
+ */