--- a/js/package.json
+++ b/js/package.json
@@ -28,6 +28,7 @@
     "gulp-sourcemaps": "^2.4.0",
     "gulp-strip-banner": "0.0.2",
     "gulp-stylelint": "^3.7.0",
+    "gulp-sync": "^0.1.4",
     "gulp-uglify": "^2.0.0",
     "gulp-uglifycss": "^1.0.6",
     "gulp-wrap": "^0.13.0",