… js/yarn.lock
… tests/js/config/karma.js