fixes #323