@skjnldsv 2017-02-02
Could you provide a before/after load test?
This is interesting! :)