@georgehrke 2017-02-02
`siege -t60s -c20 -d10 --header="Authorization:Basic YWRtaW46YWRtaW4=" 'http://nextcloud.dev/index.php/apps/calendar/'`

with bundling and minifying:
```
Lifting the server siege...
Transactions:                   1340 hits
Availability:                 100.00 %
Elapsed time:                  59.19 secs
Data transferred:              53.54 MB
Response time:                  0.87 secs
Transaction rate:              22.64 trans/sec
Throughput:                     0.90 MB/sec
Concurrency:                   19.73
Successful transactions:        1340
Failed transactions:               0
Longest transaction:            9.48
Shortest transaction:           0.00
```

without bundling and minifying:
```
Lifting the server siege...
Transactions:                   1517 hits
Availability:                 100.00 %
Elapsed time:                  59.80 secs
Data transferred:              91.28 MB
Response time:                  0.77 secs
Transaction rate:              25.37 trans/sec
Throughput:                     1.53 MB/sec
Concurrency:                   19.59
Successful transactions:        1517
Failed transactions:               0
Longest transaction:           10.73
Shortest transaction:           0.01
```