@georgehrke 2017-02-02
This makes more sense:

`siege -r6 -c20 -d10 --header="Authorization:Basic YWRtaW46YWRtaW4=" 'http://nextcloud.dev/index.php/apps/calendar/'`

with:
```
Transactions:                   8280 hits
Availability:                 100.00 %
Elapsed time:                 371.50 secs
Data transferred:             325.53 MB
Response time:                  0.80 secs
Transaction rate:              22.29 trans/sec
Throughput:                     0.88 MB/sec
Concurrency:                   17.91
Successful transactions:        8280
Failed transactions:               0
Longest transaction:           10.24
Shortest transaction:           0.00
```

without:
```
Transactions:                   9480 hits
Availability:                 100.00 %
Elapsed time:                 373.44 secs
Data transferred:             552.86 MB
Response time:                  0.70 secs
Transaction rate:              25.39 trans/sec
Throughput:                     1.48 MB/sec
Concurrency:                   17.88
Successful transactions:        9480
Failed transactions:               0
Longest transaction:           10.36
Shortest transaction:           0.00
```