@georgehrke 2017-02-02
and this is localhost,
`Data transferred: 325.53 MB` and `Data transferred: 552.86 MB` will make a big difference in real world use cases.