@georgehrke 2017-02-02
@tcitworld Bundling and Minifying the `vendor.js` and `vendor.min.js` takes a couple of seconds. That's why I decided to bundle them in a `vendor.js` instead of `app.js`. Otherwise everytime you change code and regenerate the `app.js`, you have to wait these extra couple of seconds.