@tcitworld 2017-02-12
Couldn't we get rid of the sourcemaps in production environment ?