@georgehrke 2017-02-12
> Couldn't we get rid of the sourcemaps in production environment ?

AFAIK the source maps are only loaded when you open the developer console.
Am I wrong?