@georgehrke 2017-02-12
Ok, so I'd like to keep it, because:
- it makes the package thats downloaded only once slightly larger
- it doesn't affect ordinary users
- it will make debugging in production easier 

Do you agree? :)